backend/pyproject.toml
[project]
name = "swarmmind-backend"
version = "0.1.0"
description = "Swarm Mind — LLM Agent Social Manipulation Game"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
